+++ flutter_quill_extensions/lib/presentation/embeds/toolbar/image_button/select_image_source.dart
@@ -28,7 +28,7 @@ class SelectImageSourceDialog extends StatelessWidget {
                 'Take a photo using your phone camera',
               ),
               leading: const Icon(Icons.camera),
-              enabled: !isDesktop(),
+              enabled: !isDesktop(supportWeb: false),
               onTap: () => Navigator.of(context).pop(InsertImageSource.camera),
             ),
             ListTile(