+++ flutter_quill_extensions/lib/presentation/embeds/toolbar/video_button/select_video_source.dart
@@ -28,7 +28,7 @@ class SelectVideoSourceDialog extends StatelessWidget {
                 'Record a video using your phone camera',
               ),
               leading: const Icon(Icons.camera),
-              enabled: !isDesktop(),
+              enabled: !isDesktop(supportWeb: false),
               onTap: () => Navigator.of(context).pop(InsertVideoSource.camera),
             ),
             ListTile(