+++ lib/src/widgets/cursor.dart
@@ -291,7 +291,7 @@ class CursorPainter {
 
     final caretHeight = editable!.getFullHeightForCaret(position);
     if (caretHeight != null) {
-      if (isAppleOS()) {
+      if (isAppleOS(supportWeb: true)) {
         // Center the caret vertically along the text.
         caretRect = Rect.fromLTWH(
           caretRect.left,